+++ Tools/bench_test/px4bench/missions.py
@@ -1,18 +1,4 @@
-#!/usr/bin/env python3
-"""
-Mission upload/download/clear torture test for on-bench PX4 hardware.
-
-v1.18 risk area: dataman and the mission shared-state path were reworked and
-a brand-new mutex now guards mission shared state. A regression there shows
-up as a SILENT HANG (a MISSION_REQUEST that never arrives, a MISSION_ACK that
-never comes, a download that stalls mid-list) or as a corrupted round-trip.
-This script repeatedly uploads a large mission, reads it back and compares it
-item-by-item, then clears it, with a hard timeout on every wait so a stall
-becomes a named FAIL naming the stalled step and the seq reached.
-
-With a second connection given, iterations alternate between the two links
-(iteration parity picks the link) to exercise the new mission shared-state
-mutex from two MAVLink channels.
+"""Mission protocol helpers: item generation, upload/download/compare/clear.
 
 Mission protocol (verified in src/modules/mavlink/mavlink_mission.cpp):
   upload:   MISSION_COUNT -> per-item MISSION_REQUEST_INT / MISSION_REQUEST
@@ -21,25 +7,15 @@ Mission protocol (verified in src/modules/mavlink/mavlink_mission.cpp):
             MISSION_REQUEST_INT -> we reply, then send MISSION_ACK(ACCEPTED)
   clear:    MISSION_CLEAR_ALL -> MISSION_ACK(ACCEPTED)
 
-Usage:
-    mission_torture.py CONNECTION [CONNECTION2] [-b BAUD]
-                       [--baudrate2 BAUD2] [--iterations N] [--items K]
+The autopilot may request items with either MISSION_REQUEST_INT or the
+deprecated MISSION_REQUEST; both are answered with MISSION_ITEM_INT, and
+duplicate or out-of-order requests are answered as asked.
 """
 
-import argparse
 import math
-import os
-import sys
 import time
 
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-import px4bench
-from px4bench import mavutil
-
-DEFAULT_ITERATIONS = 10
-DEFAULT_ITEMS = 220        # <= smallest CONFIG_NUM_MISSION_ITMES_SUPPORTED (500)
-DEFAULT_BAUD2 = 57600
+from pymavlink import mavutil
 
 MAV_CMD_NAV_WAYPOINT = mavutil.mavlink.MAV_CMD_NAV_WAYPOINT
 MAV_FRAME = mavutil.mavlink.MAV_FRAME_GLOBAL_RELATIVE_ALT_INT
@@ -49,10 +25,13 @@ MAV_MISSION_ACCEPTED = mavutil.mavlink.MAV_MISSION_ACCEPTED
 BASE_LAT = 47.397742
 BASE_LON = 8.545594
 
-# Per-message wait cap and per-iteration transaction deadline.
 STEP_TIMEOUT_S = 5.0
 TRANSACTION_DEADLINE_S = 60.0
 COUNT_RETRANSMITS = 3
+
+UPLOAD_TYPES = ['MISSION_REQUEST_INT', 'MISSION_REQUEST', 'MISSION_ACK']
+DOWNLOAD_COUNT_TYPES = ['MISSION_COUNT', 'MISSION_ACK']
+DOWNLOAD_ITEM_TYPES = ['MISSION_ITEM_INT', 'MISSION_ITEM', 'MISSION_ACK']
 
 
 def transaction_deadline(num_items):
@@ -64,15 +43,6 @@ def transaction_deadline(num_items):
     """
     return max(TRANSACTION_DEADLINE_S, 0.5 * num_items)
 
-# Messages we care about; anything else (PARAM_VALUE, HEARTBEAT, ...) is drained.
-UPLOAD_TYPES = ['MISSION_REQUEST_INT', 'MISSION_REQUEST', 'MISSION_ACK']
-DOWNLOAD_COUNT_TYPES = ['MISSION_COUNT', 'MISSION_ACK']
-DOWNLOAD_ITEM_TYPES = ['MISSION_ITEM_INT', 'MISSION_ITEM', 'MISSION_ACK']
-
-
-# ---------------------------------------------------------------------------
-# Deterministic mission generation
-# ---------------------------------------------------------------------------
 
 class Item:
     """A single generated waypoint; comparison uses the fields we round-trip."""
@@ -102,10 +72,6 @@ def generate_mission(iteration, k):
         items.append(Item(seq, lat_int, lon_int, alt, 1 if seq == 0 else 0))
     return items
 
-
-# ---------------------------------------------------------------------------
-# Upload
-# ---------------------------------------------------------------------------
 
 def send_item(mav, item):
     mav.mav.mission_item_int_send(
@@ -170,10 +136,6 @@ def upload_mission(report, mav, items, iteration):
                 transaction_deadline(len(items)), seq_reached))
 
 
-# ---------------------------------------------------------------------------
-# Download + compare
-# ---------------------------------------------------------------------------
-
 def download_mission(mav):
     """Request and collect the mission. Returns (items_or_None, duration, detail)."""
     start = time.monotonic()
@@ -182,7 +144,6 @@ def download_mission(mav):
     mav.mav.mission_request_list_send(
         mav.target_system, mav.target_component, MAV_MISSION_TYPE_MISSION)
 
-    # Wait for MISSION_COUNT.
     count = None
     while time.monotonic() < deadline:
         m = mav.recv_match(type=DOWNLOAD_COUNT_TYPES, blocking=True, timeout=STEP_TIMEOUT_S)
@@ -255,10 +216,6 @@ def compare_mission(expected, downloaded):
     return True, 'all {} items match'.format(len(expected))
 
 
-# ---------------------------------------------------------------------------
-# Clear
-# ---------------------------------------------------------------------------
-
 def clear_mission(mav):
     """MISSION_CLEAR_ALL -> MISSION_ACK(ACCEPTED). Returns (ok, detail)."""
     deadline = time.monotonic() + TRANSACTION_DEADLINE_S
@@ -298,100 +255,3 @@ def verify_cleared(mav):
                 return True, 'empty mission (ACK) after clear'
             return False, 'verify-clear got MISSION_ACK type {}'.format(m.type)
     return False, 'verify-clear transaction deadline hit'
-
-
-# ---------------------------------------------------------------------------
-# main
-# ---------------------------------------------------------------------------
-
-def run_iteration(report, mav, iteration, items_k, link_label):
-    """One full upload/download/compare/clear cycle. Returns True on full pass."""
-    expected = generate_mission(iteration, items_k)
-
-    up_ok, up_dur, up_detail = upload_mission(report, mav, expected, iteration)
-    if not report.check('iter{}_upload'.format(iteration), up_ok,
-                        '[{}] {} ({:.1f}s)'.format(link_label, up_detail, up_dur)):
-        return False
-
-    downloaded, dl_dur, dl_detail = download_mission(mav)
-    if downloaded is None:
-        report.fail('iter{}_download'.format(iteration),
-                    '[{}] {} ({:.1f}s)'.format(link_label, dl_detail, dl_dur))
-        return False
-
-    cmp_ok, cmp_detail = compare_mission(expected, downloaded)
-    if not report.check('iter{}_compare'.format(iteration), cmp_ok,
-                        '[{}] {}'.format(link_label, cmp_detail)):
-        return False
-
-    clr_ok, clr_detail = clear_mission(mav)
-    if not report.check('iter{}_clear'.format(iteration), clr_ok,
-                        '[{}] {}'.format(link_label, clr_detail)):
-        return False
-
-    vfy_ok, vfy_detail = verify_cleared(mav)
-    report.check('iter{}_verify_clear'.format(iteration), vfy_ok,
-                 '[{}] {}'.format(link_label, vfy_detail))
-
-    if up_ok and downloaded is not None and cmp_ok and clr_ok and vfy_ok:
-        report.info('iter {} PASS [{}] upload {:.1f}s download {:.1f}s'.format(
-            iteration, link_label, up_dur, dl_dur))
-        return True
-    return False
-
-
-def main():
-    parser = argparse.ArgumentParser(
-        description=__doc__,
-        formatter_class=argparse.RawDescriptionHelpFormatter)
-    px4bench.add_connection_args(parser, dual_link=True)
-    parser.add_argument('--baudrate2', type=int, default=DEFAULT_BAUD2,
-                        help='baud rate for the second connection (default: %(default)s)')
-    parser.add_argument('--iterations', type=int, default=DEFAULT_ITERATIONS,
-                        help='upload/download/clear iterations (default: %(default)s)')
-    parser.add_argument('--items', type=int, default=DEFAULT_ITEMS,
-                        help='mission items per iteration (default: %(default)s)')
-    args = parser.parse_args()
-
-    report = px4bench.Reporter('mission_torture')
-
-    try:
-        mav1 = px4bench.connect(args.connection, baud=args.baudrate,
-                                timeout=args.connect_timeout)
-    except (TimeoutError, OSError) as e:
-        report.fail('connect', 'link 1 {}: {}'.format(args.connection, e))
-        sys.exit(report.finish())
-    report.info('link 1 connected to system {} component {}'.format(
-        mav1.target_system, mav1.target_component))
-
-    mav2 = None
-    if args.connection2:
-        try:
-            mav2 = px4bench.connect(args.connection2, baud=args.baudrate2,
-                                    timeout=args.connect_timeout)
-            report.info('link 2 connected to system {} component {}'.format(
-                mav2.target_system, mav2.target_component))
-        except (TimeoutError, OSError) as e:
-            report.fail('connect', 'link 2 {}: {}'.format(args.connection2, e))
-            sys.exit(report.finish())
-
-    try:
-        for iteration in range(args.iterations):
-            if mav2 is not None and (iteration % 2 == 1):
-                mav, label = mav2, 'link2'
-            else:
-                mav, label = mav1, 'link1'
-            run_iteration(report, mav, iteration, args.items, label)
-    finally:
-        for m in (mav1, mav2):
-            try:
-                if m is not None:
-                    m.close()
-            except Exception:
-                pass
-
-    sys.exit(report.finish())
-
-
-if __name__ == '__main__':
-    main()
