+++ Tools/bench_test/pyproject.toml
@@ -1,0 +1,24 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "px4bench"
+version = "0.1.0"
+description = "PX4 bench-test suite: hang-visible release qualification on real NuttX hardware"
+readme = "README.md"
+license = { text = "BSD-3-Clause" }
+requires-python = ">=3.8"
+dependencies = [
+    "pymavlink>=2.4.40",
+    "pyserial",
+]
+
+[project.optional-dependencies]
+ulog = ["pyulog"]
+
+[project.urls]
+Homepage = "https://github.com/PX4/PX4-Autopilot"
+
+[tool.setuptools]
+packages = ["px4bench"]
